lzl_apia/requirements.txt
redis==3.5.3
django_filter==2.4.0
qcloudsms_py==0.1.4
requests==2.26.0
Django==2.2.2
djangorestframework_jwt==1.11.0
beautifulsoup4==4.9.3
djangorestframework==3.12.4
PyJWT==1.7.1
